pyproject.toml
[project]
name = "site-survey-app"
version = "0.1.0"
description = "Offline-first site survey application with CRDT-based synchronization"
authors = [
    {name = "Keith", email = "keith@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "briefcase>=0.3.17",
    "flask>=2.3.0",
    "flask-sqlalchemy>=3.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "psycopg2-binary>=2.9.0",
    "toga>=0.4.0",
    "pillow>=10.0.0",
    "appdirs>=1.4.4",
    # Testing dependencies
    "pytest>=7.4.0",
    "pytest-flask>=1.3.0",
    "pytest-cov>=4.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "src"]

[tool.briefcase]
project_name = "Site Survey App"
bundle = "com.keith.surveyapp"
version = "0.1.0"
url = "https://example.com"
license = "BSD license"
author = "Keith"
author_email = "keith@example.com"
output_format = "app"

[tool.briefcase.app.survey_app]
formal_name = "Site Survey App"
description = "Offline-first site survey application"
icon = "icon"
sources = ["src/survey_app"]
requires = [
    "toga>=0.4.0",
    "pillow>=10.0.0",
    "requests>=2.31.0"
]